Move manual setup out of init.py
